% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R, R/viromeBrowser.R
\docType{package}
\name{viromeBrowser}
\alias{viromeBrowser}
\title{Launch the Virome Browser}
\usage{
viromeBrowser(
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port", 3838)
)
}
\arguments{
\item{host}{The host ip address, 127.0.0.1 or "localhost" by default.}

\item{port}{The host port, 3838 by default.}
}
\description{
An application for viewing virome annotation files

The viromeBrowser package provides a single function to launch the virome browser shiny app
}
\section{viromeBrowser functions}{

viromeBrowser()
}

